swarm-host: Rework event flow of the system
Remove subscription-based model from event stream and instead
return the interface handle and its event stream together which
`Overseer` can then poll for events. Split the socket into two
and send the writer half along with `PeerConnected` event to `Overseer`.
In its current for this is way too leaky and proper fix for it has
to be implemented some time in the future.