swarm-host: Add support for generic packet filters
Business logic of how to handle the traffic interfaces can be now written
in Python. The code requires a lot of cleaning to abstract `PyO3` behind
some interface so `MessageFilter` doesn't have to concern itself with the
details of the packet filter executor but this is a good first step.