swarm-host: Add initial Proof-of-Concept
Implement simple Python framework which allows users to start
`swarm-host`, modify its behavior, spawn Substrate nodes and connect
them to `swarm-host` and most importantly, install custom protocol
filters written in Python which the `swarm-host` then calls upon
receiving a notification over the specified protocol.

These custom filters receive the source and destination interfaces and
peer IDs and the received notification which can be anything based on
what the protocol is, then decode the notification and make decision
based on the source, destination, content of the notification and any
other auxiliary information they might stored in the context whether to
forward the received notification and to whom.

The code is currently very ugly, inefficient and leaky but those can be
fixed with relative ease.