swarm-host: Add PoC for request-response support
The basic idea is simple: filters share a data storage from which they
can query data upon receiving a request. If the requested data is not
available, the request is forwarded to one of the honest nodes which then
responds to it and when the response is received by the filter, it stores
the data in the shared storage and responds to the original request.

The code needs a lot of clean up and it's far from complete in terms of
needed features but this commit provides the foundation for future features.