framework: Add simple block request handler
Honestly speaking the code could not be much worse but it works and
it's easy to improve now that there is something that actually handles
inbound block requests.

Each filter must have a block request handler if the filter forwards
any block announcements forward as any connected may send the filter
a block request for the received announcement.