backend/substrate: Don't register peers on `ConnectionEstablished`
Interfaces are able to connect to each other even if they're not going
to open substreams to each other so if `ConnectionEstablished` event is
used to register peers to `Overseer`, `Filter` will contain other
interfaces as peers which is not desirable.

Register peers to `Overseer` only when `/sup/block-announces/1` is
opened and register all protocols after that as connection upgrades.

This relies on the assumption from Substrate that block announces
substream is the first stream that is opened.