swarm-host: Add preinit hook for interface creation
Give user the ability to call a custom executor code before the interface
is created and use the return value of that executor call to initialize
the interface.

This is useful for situations where the networking backend may need some
runtime data to finalize its initialization (such as a custom handshake)
to be fully operational.