executor: Introduce `ExecutorEvent`
Currently the filters are purely input-drive which is convenient for
some tests but other tests might require that the filter can be polled
independently of any received event.

This commit lays the foundation for a poll-based approach for executors
where the programmer can specify that the executor can be called
independelty of any received event and allows, for example, the filter
to send a request that is not prompted by any received response.

The idea is that an event is received and it's injected to the correct
filter but that filter in itself doesn't return anything but if the call
would result in some action that should be performed by `swarm-host`,
the call result is received by calling `poll()` after the first filter
call finishes.