swarm-host: Give filters the ability to establish outbound connections
When `discover_peer()` is called, the filter has the ability to mark
the peer as pending and push an event to pending events which instructs
the `Filter` to attempt to establish an outbound connection to peer.

`Notifications` may grumble that the connection is already active because
the filter gets the discovery and connection events at a delay but this
is not a fatal error even though there was a `debug_assert!()` because
this would be incorrect behavior from `Peerset` which works synchronously
but not from a filter which works asynchronously.