framework: Covert flash attack
After Sybil nodes stop forwarding messages, `SyncingEngine` evicts some of
them but the network is not able to recover due to the validator being
swarmed with Sybil nodes.

The exact result of the test depends on how the Substrate nodes get peered
when they're booting. If one of them is able to establish a connection to
the validator, the network can recover because idle peers get evicted.