framework: Cold boot attack
Depending on how on how the peering during boot goes, the network either
completely halts or there may be 1 or 2 nodes which get peered with the
validator and are able to recover from the attack.

Other nodes which are not so lucky evict the idle peers and their peer
count stays between 5-7 but best block is stuck.